+++ meta-memfault/recipes-memfault/memfaultd/files/memfaultd/src/cli/memfault_core_handler/core_elf_memfault_note.rs
@@ -1,6 +1,16 @@
 //
 // Copyright (c) Memfault, Inc.
 // See License.txt for details
+//! Utilities and data types for writing Memfault-specific ELF notes to a core dump file.
+//!
+//! Currently we write two notes:
+//!
+//! 1. A note containing metadata about the core dump. This note is written by the
+//!   `CoreHandler` whenever it receives a core dump. It contains information about the device,
+//!    that will be used to associate the core dump with a device in the Memfault cloud.
+//! 2. A note containing debug data about the core dump. Currently this note only contains
+//!    logs written during the coredump capture process. These logs are used by Memfault to debug
+//!    issues with coredump capture.
 use std::time::SystemTime;
 
 use crate::build_info::VERSION;
@@ -8,13 +18,19 @@ use crate::config::CoredumpCaptureStrategy;
 
 use ciborium::{cbor, into_writer};
 use eyre::Result;
+use serde::Serialize;
 
 use super::core_elf_note::build_elf_note;
 
 const NOTE_NAME: &str = "Memfault\0";
-const NOTE_TYPE: u32 = 0x4154454d;
+const METADATA_NOTE_TYPE: u32 = 0x4154454d;
+const DEBUG_DATA_NOTE_TYPE: u32 = 0x4154454e;
 const MEMFAULT_CORE_ELF_METADATA_SCHEMA_VERSION_V1: u32 = 1;
+const MEMFAULT_CORE_ELF_DEBUG_DATA_SCHEMA_VERSION_V1: u32 = 1;
 
+/// Map of keys used in the Memfault core ELF metadata note.
+///
+/// Integer keys are used here instead of strings to reduce the size of the note.
 enum MemfaultCoreElfMetadataKey {
     SchemaVersion = 1,
     LinuxSdkVersion = 2,
@@ -27,6 +43,7 @@ enum MemfaultCoreElfMetadataKey {
     CaptureStrategy = 9,
 }
 
+/// Metadata about a core dump.
 #[derive(Debug)]
 pub struct CoredumpMetadata {
     pub device_id: String,
@@ -57,6 +74,9 @@ impl CoredumpMetadata {
     }
 }
 
+/// Serialize a `CoredumpMetadata` struct as a CBOR map.
+///
+/// This CBOR map uses integer keys instead of strings to reduce the size of the note.
 pub fn serialize_metadata_as_map(metadata: &CoredumpMetadata) -> Result<Vec<u8>> {
     let cbor_val = cbor!({
         MemfaultCoreElfMetadataKey::SchemaVersion as u32 => MEMFAULT_CORE_ELF_METADATA_SCHEMA_VERSION_V1,
@@ -76,10 +96,40 @@ pub fn serialize_metadata_as_map(metadata: &CoredumpMetadata) -> Result<Vec<u8>>
     Ok(buffer)
 }
 
-pub fn write_memfault_note(metadata: &CoredumpMetadata) -> Result<Vec<u8>> {
+/// Write a core ELF note containing metadata about a core dump.
+///
+/// This note is written by the `CoreHandler` whenever it receives a core dump. It contains
+/// information about the device, that will be used to associate the core dump with a device in the
+/// Memfault cloud.
+pub fn write_memfault_metadata_note(metadata: &CoredumpMetadata) -> Result<Vec<u8>> {
     let description_buffer = serialize_metadata_as_map(metadata)?;
 
-    build_elf_note(NOTE_NAME, &description_buffer, NOTE_TYPE)
+    build_elf_note(NOTE_NAME, &description_buffer, METADATA_NOTE_TYPE)
+}
+
+/// A note containing a list of errors that occurred during coredump capture.
+///
+/// This note is written by the `CoreHandlerLogWrapper` when it receives an error or warning log.
+/// These logs will help us debug issues with coredump capture.
+#[derive(Debug, Serialize)]
+pub struct CoredumpDebugData {
+    pub schema_version: u32,
+    pub capture_logs: Vec<String>,
+}
+
+/// Write a core ELF note containing debug data about the coredump capture process.
+///
+/// See `CoredumpDebugData` for more information.
+pub fn write_memfault_debug_data_note(errors: Vec<String>) -> Result<Vec<u8>> {
+    let coredump_capture_logs = CoredumpDebugData {
+        schema_version: MEMFAULT_CORE_ELF_DEBUG_DATA_SCHEMA_VERSION_V1,
+        capture_logs: errors,
+    };
+
+    let mut buffer = Vec::new();
+    into_writer(&coredump_capture_logs, &mut buffer)?;
+
+    build_elf_note(NOTE_NAME, &buffer, DEBUG_DATA_NOTE_TYPE)
 }
 
 #[cfg(test)]
@@ -116,5 +166,24 @@ mod test {
         set_snapshot_suffix!("{}", test_name);
         insta::assert_debug_snapshot!(deser_map);
         assert_eq!(map.len(), expected_size);
+    }
+
+    #[test]
+    fn serialize_debug_data() {
+        let capture_logs = CoredumpDebugData {
+            schema_version: MEMFAULT_CORE_ELF_DEBUG_DATA_SCHEMA_VERSION_V1,
+            capture_logs: vec![
+                "Error 1".to_string(),
+                "Error 2".to_string(),
+                "Error 3".to_string(),
+            ],
+        };
+
+        let mut capture_logs_buffer = Vec::new();
+        into_writer(&capture_logs, &mut capture_logs_buffer).unwrap();
+
+        let deser_capture_logs: Value = from_reader(capture_logs_buffer.as_slice()).unwrap();
+
+        insta::assert_debug_snapshot!(deser_capture_logs);
     }
 }