+++ test_scripts/test_memfaultd_enable_data_collection.py
@@ -2,6 +2,7 @@
 # Copyright (c) Memfault, Inc.
 # See License.txt for details
 import time
+from typing import cast
 
 import pexpect
 import pytest
@@ -49,7 +50,7 @@ def test_start(qemu: QEMU) -> None:
     output = b""
     while True:
         try:
-            output += qemu.child().read_nonblocking(size=1024, timeout=0.5)
+            output += qemu.child().read_nonblocking(size=1024, timeout=cast(int, 0.5))
         except pexpect.TIMEOUT:
             break
     assert b"memfaultd.service: Scheduled restart job" not in output
@@ -86,7 +87,7 @@ def test_via_memfaultctl(qemu: QEMU) -> None:
     output = b""
     while True:
         try:
-            output += qemu.child().read_nonblocking(size=1024, timeout=0.5)
+            output += qemu.child().read_nonblocking(size=1024, timeout=cast(int, 0.5))
         except pexpect.TIMEOUT:
             break
     assert b"memfaultd.service: Scheduled restart job" not in output