+++ test_scripts/test_memfaultd_enable_dev_mode.py
@@ -1,6 +1,8 @@
 #
 # Copyright (c) Memfault, Inc.
 # See License.txt for details
+from typing import cast
+
 import pexpect
 
 from .qemu import QEMU
@@ -39,7 +41,7 @@ def test_start(qemu: QEMU) -> None:
     output = b""
     while True:
         try:
-            output += qemu.child().read_nonblocking(size=1024, timeout=0.5)
+            output += qemu.child().read_nonblocking(size=1024, timeout=cast(int, 0.5))
         except pexpect.TIMEOUT:
             break
     assert b"memfaultd.service: Scheduled restart job" not in output